# src/main/swagger/vision.yaml
swagger: '2.0'
info:
  version: 0.1-SNAPSHOT
  title: Vision
basePath: /rest
...
paths:
...
  /user/login:
    post:
      tags:
        - user
      summary: Login and create a session
      description: ''
      operationId: login
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/LoginRequest'
      responses:
        default:
          description: successful operation
...
  /user/profile:
    get:
      tags:
        - user
      summary: Retrieve user's profile
      description: ''
      operationId: get
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User information'
    post:
      tags:
        - user
      summary: Save user's profile
      description: ''
      operationId: save
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Profile modification request'
      responses:
        default:
          description: successful operation
...
definitions:
...
  LoginRequest:
    type: object
    required:
      - name
      - password
    properties:
      name:
        type: string
        description: username
      password:
        type: string
        description: password
...